rename statsd client files
