+++ proxy/statsdclient/client.go
@@ -1,5 +1,14 @@
 package statsdclient
 
+import GoMetricStatsdClient "github.com/GoMetric/go-statsd-client"
+
+func NewGoMetricClient(
+	statsdHost string,
+	statsdPort int,
+) StatsdClientInterface {
+	return GoMetricStatsdClient.NewClient(statsdHost, statsdPort)
+}
+
 type StatsdClientInterface interface {
 	Open()
 	Close()
