+++ proxy/statsdclient/client_test.go
