--- proxy/statsdclient/gometric.go
@@ -1,10 +1,0 @@
-package statsdclient
-
-import GoMetricStatsdClient "github.com/GoMetric/go-statsd-client"
-
-func NewGoMetricClient(
-	statsdHost string,
-	statsdPort int,
-) StatsdClientInterface {
-	return GoMetricStatsdClient.NewClient(statsdHost, statsdPort)
-}
